Working multi-server follow test after refactoring
