+++ src/middleware/packages/ldp/services/container/actions/patch.js
@@ -70,42 +70,13 @@ module.exports = {
                   // we need to import the remote resource
                   this.logger.info('IMPORTING ' + insUri);
                   try {
-                    const response = await fetch(insUri, { headers: { Accept: MIME_TYPES.JSON } });
-                    const resource = await response.json();
+                    await ctx.call('mirror.resource.store', {
+                      resourceUri: insUri,
+                      keepInSync: true,
+                      webId
+                    })
 
-                    // TODO if JSON is not available, try Turtle
-
-                    await ctx.call('triplestore.insert', {
-                      resource,
-                      contentType: MIME_TYPES.JSON,
-                      webId,
-                      graphName: this.settings.mirrorGraphName
-                    });
-
-                    // try {
-                    //   newResource = JSON.parse(newResource);
-                    //   newResource = await ctx.call('jsonld.toRDF', { input: newResource });
-                    //   console.log('new resource', newResource)
-                    // } catch(e) {
-                    //   // Ignore if this is not a JSON
-                    // }
-                    //
-                    // const prefixes = [...newResource.matchAll(regexPrefix)];
-                    //
-                    // let turtleToSparql = '';
-                    // for (const pref of prefixes) {
-                    //   turtleToSparql += 'PREFIX ' + pref[1] + '\n';
-                    // }
-                    // turtleToSparql += `INSERT DATA { GRAPH <${this.settings.mirrorGraphName}> { \n`;
-                    // turtleToSparql += newResource.replace(regexPrefix, '');
-                    // turtleToSparql += `<${insUri}> <http://semapps.org/ns/core#singleMirroredResource> <${
-                    //   new URL(insUri).origin
-                    // }> .`;
-                    // turtleToSparql += '} }';
-                    //
-                    // await ctx.call('triplestore.update', { query: turtleToSparql });
-
-                    // now if the import went well, we can retry the attach
+                    // Now if the import went well, we can retry the attach
                     await ctx.call('ldp.container.attach', { containerUri, resourceUri: insUri });
                   } catch (e) {
                     this.logger.warn('ERROR while IMPORTING ' + insUri + ' : ' + e.message);
@@ -122,29 +93,12 @@ module.exports = {
               try {
                 await ctx.call('ldp.container.detach', { containerUri, resourceUri: delUri });
 
-                // if the resource is not attached to any container anymore, it must be deleted.
-
-                let remaining = await ctx.call('triplestore.query', {
-                  query: `SELECT (COUNT (?s) as ?count) WHERE { ?s <http://www.w3.org/ns/ldp#contains> <${delUri}> }`
-                });
-                remaining = Number(remaining[0].count.value);
-                if (remaining === 0) {
-                  const isM = isMirror(delUri, this.settings.baseUrl);
-
-                  await ctx.call('triplestore.update', {
-                    query: `
-                    DELETE
-                    WHERE { 
-                      ${isM ? 'GRAPH <' + this.settings.mirrorGraphName + '> {' : ''}
-                      <${delUri}> ?p1 ?o1 . ${isM ? '}' : ''}
-                    }
-                  `,
-                    webId: 'system'
-                  });
-
-                  ctx.call('triplestore.deleteOrphanBlankNodes', {
-                    graphName: isM ? this.settings.mirrorGraphName : undefined
-                  });
+                // If the mirrored resource is not attached to any container anymore, it must be deleted.
+                const containers = await ctx.call('ldp.resource.getContainers', { resourceUri: delUri });
+                if (containers.length === 0 && isMirror(delUri, this.settings.baseUrl)) {
+                  await ctx.call('mirror.resource.delete', {
+                    resourceUri: delUri
+                  })
 
                   ctx.emit(
                     'ldp.resource.deletedSingleMirror',
@@ -153,7 +107,8 @@ module.exports = {
                   );
                 }
               } catch (e) {
-                // fail silently
+                // Fail silently
+                this.logger.warn(`Error when detaching ${delUri} from ${containerUri}: ${e.message}`);
               }
             }
         }
@@ -164,9 +119,7 @@ module.exports = {
 
       ctx.emit(
         'ldp.container.patched',
-        {
-          containerUri
-        },
+        { containerUri },
         { meta: { webId: null, dataset: null } }
       );
     }
