+++ src/middleware/packages/ldp/services/registry/defaultOptions.js
@@ -1,7 +1,6 @@
 module.exports = {
   accept: 'text/turtle',
   jsonContext: null,
-  queryDepth: 0,
   dereference: [],
   readOnly: false,
   excludeFromMirror: false,
