+++ src/middleware/packages/ldp/services/resource/index.js
@@ -10,11 +10,8 @@ const uploadAction = require('./actions/upload');
 const headAction = require('./actions/head');
 const methods = require('./methods');
 
-const Schedule = require('moleculer-schedule');
-
 module.exports = {
   name: 'ldp.resource',
-  mixins: [Schedule],
   settings: {
     baseUrl: null,
     ontologies: [],
@@ -58,11 +55,5 @@ module.exports = {
       }
     }
   },
-  methods,
-  jobs: [
-    {
-      rule: '0 * * * *',
-      handler: 'updateSingleMirroredResources'
-    }
-  ]
+  methods
 };
