+++ src/middleware/packages/ldp/utils.js
@@ -25,26 +25,6 @@ const isMirror = (resourceUri, baseUrl) => {
   return !urlJoin(resourceUri, '/').startsWith(baseUrl);
 };
 
-const buildBlankNodesQuery = depth => {
-  let construct = '',
-    where = '';
-  if (depth > 0) {
-    for (let i = depth; i >= 1; i--) {
-      construct += `
-        ?o${i} ?p${i + 1} ?o${i + 1} .
-      `;
-      where = `
-        OPTIONAL {
-          FILTER((isBLANK(?o${i}))) .
-          ?o${i} ?p${i + 1} ?o${i + 1} .
-          ${where}
-        }
-      `;
-    }
-  }
-  return { construct, where };
-};
-
 // Transform ['ont:predicate1/ont:predicate2'] to ['ont:predicate1', 'ont:predicate1/ont:predicate2']
 const extractNodes = predicates => {
   let nodes = [];
@@ -179,7 +159,6 @@ const defaultToArray = value => (!value ? undefined : Array.isArray(value) ? val
 const delay = t => new Promise(resolve => setTimeout(resolve, t));
 
 module.exports = {
-  buildBlankNodesQuery,
   buildDereferenceQuery,
   buildFiltersQuery,
   getPrefixRdf,
