+++ website/docs/middleware/ldp/index.md
@@ -86,7 +86,6 @@ The following options can be set for each container, or they can be set in the `
 | `excludeFromMirror`       | `Boolean`              | false         | If true, other servers will not be able to [mirror](../mirror) this container.                                                                         |
 | `jsonContext`             | `Any`                  |               | JSON context to use to format results                                                                                                                  |
 | `dereference`             | `Array`                |               | Blank nodes to dereference, prefixed with their ontology. You can define sub-predicates separated by `/`                                               |
-| `queryDepth`              | `Integer`              | 0             | Depth of blank nodes to dereference (Deprecated. Will be removed in a future minor release.)                                                           |
 | `permissions`             | `Object` or `Function` |               | If the WebACL service is activated, permissions of the container itself                                                                                |
 | `newResourcesPermissions` | `Object` or `Function` |               | If the WebACL service is activated, permissions for new resources. [See the docs here](../webacl/index.md#default-permissions-for-new-resources)       |
 | `readOnly`                | `Boolean`              | false         | Do not set `POST`, `PATCH`, `PUT` and `DELETE` routes for the container and its resources                                                              |
